Survey can now save and store answers from user in the database, as well as the evaluator can now start and end the survey

+++ back-end/cybersecurity-awareness/src/main/java/ntnu/master/assessment/cybersecurityawareness/persistance/entity/Answer.java
@@ -21,7 +21,7 @@ public class Answer {
     private int questionId;
 
     @Column(name = "answer")
-    private double answer;
+    private int answer;
 
     @Column(name = "Organization_ID")
     private int organizationId;
